+++ newrelic/config.py
@@ -2,7 +2,6 @@ import os
 import sys
 import logging
 import traceback
-import warnings
 
 try:
     import ConfigParser
@@ -766,16 +765,10 @@ def translate_deprecated_settings(settings, cached_settings):
                 'request parameters, add "request.parameters.*" to '
                 'attributes.exclude.')
 
-    # Log a DeprecationWarning for customers who have disabled SSL. Disabling
-    # SSL will be disallowed in a future agent version.
     if not settings.ssl:
-        msg = ('Disabling SSL will be disallowed in a future agent version. '
-               'Please enable ssl by removing ssl=false from your New Relic '
-               'configuration file or by removing the NEW_RELIC_SSL '
-               'environment variable from the application environment.')
-
-        warnings.warn(msg, DeprecationWarning, stacklevel=2)
-        _logger.warning(msg)
+        settings.ssl = True
+        _logger.info('Ignoring deprecated setting: ssl. Enabling ssl is now '
+                'mandatory. Setting ssl=true.')
 
     return settings
 
@@ -791,10 +784,6 @@ def apply_local_high_security_mode_setting(settings):
     log_template = ('Overriding setting for %r because High '
                     'Security Mode has been activated. The original '
                     'setting was %r. The new setting is %r.')
-
-    if not settings.ssl:
-        settings.ssl = True
-        _logger.info(log_template, 'ssl', False, True)
 
     # capture_params is a deprecated setting for users, and has three
     # possible values: