+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 unreleased
 ----------
 
+- Support for AIOHTTP 3
+
+  AIOHTTP major version 3 is now supported by the New Relic Python agent.
+
 2.104.0 (2018-02-20)
 --------------------